pyproject.toml
[project]
name = "bitbank-binance-bot"
version = "0.1.0"
description = "Follow-trading bot using bitbank.nz signals on Binance"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-binance>=1.0.19",
    "pandas>=2.0",
    "python-dotenv>=1.0",
    "scipy>=1.11",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[project.scripts]
bitbank-bot = "bot:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
